refactor: 게시글과 글쓰기 관련 API 역할 분리와 타입 분리 리팩토링
