+++ src/lib/api/post.client.ts
@@ -1,4 +1,25 @@
 import { client } from "@/lib/fetch/client";
+import type { CreatePostPayload, CreatePostResponse } from "@/types/editor";
+
+// 게시글 작성
+export const createPost = async (
+  payload: CreatePostPayload,
+): Promise<CreatePostResponse> => {
+  return await client<CreatePostResponse>("/posts", {
+    method: "POST",
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify(payload),
+    needAuth: true,
+  });
+};
+
+// 게시글 삭제
+export async function deletePost(postId: number): Promise<void> {
+  await client(`/posts/${postId}`, {
+    method: "DELETE",
+    needAuth: true,
+  });
+}
 
 // 게시글 좋아요 여부 조회
 export const getPostLikeStatus = async (postId: number): Promise<boolean> => {
