+++ src/lib/api/posts.ts
@@ -2,29 +2,6 @@ import { client } from "@/lib/fetch/client";
 import { QueryParams } from "../fetch/util/objectToQueryString";
 import { PostDetailResponse } from "@/types/post";
 
-type CreatePostResponse = {
-  category: string;
-  postId: number;
-};
-
-// 게시글 작성
-export async function createPost(payload: {
-  title: string;
-  content: string;
-  category: string;
-  tag?: string;
-  postImageList?: string[];
-  thumbnailImage?: string | null;
-}): Promise<CreatePostResponse> {
-  return await client<CreatePostResponse>("/posts", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(payload),
-  });
-}
-
 export interface FetchPostsParams extends QueryParams {
   category: string;
   tag?: string;
@@ -70,11 +47,3 @@ export const fetchSearchPosts = async ({
     },
   });
 };
-
-// 게시글 삭제
-export async function deletePost(postId: number): Promise<void> {
-  await client(`/posts/${postId}`, {
-    method: "DELETE",
-    needAuth: true,
-  });
-}
